pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xai-shap-framework"
version = "1.0.0"
description = "A comprehensive framework for Visual Analytics of Black-Box Models using SHAP for Responsible and Explainable AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "XAI-SHAP Framework Team"}
]
keywords = [
    "explainable-ai",
    "xai",
    "shap",
    "machine-learning",
    "interpretability",
    "fairness",
    "visual-analytics",
    "responsible-ai"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "xgboost>=1.5.0",
    "shap>=0.41.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "streamlit>=1.10.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
]
full = [
    "lightgbm>=3.3.0",
    "tensorflow>=2.8.0",
    "torch>=1.10.0",
    "aif360>=0.5.0",
    "lime>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/xai-shap-framework"
Documentation = "https://github.com/xai-shap-framework/docs"
Repository = "https://github.com/xai-shap-framework"
Issues = "https://github.com/xai-shap-framework/issues"

[project.scripts]
xai-shap = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
